Use suspense for loading state
